requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
numpy>=1.24.0
mss>=9.0.0
Pillow>=10.0.0
pyzmq>=25.0.0
pyobjc-framework-Vision>=10.0
pyobjc-framework-Quartz>=10.0
